Modify debug prints
Puts a # in front of the ball x, y printout for every frame. Artifacts are still not resolved. Suspect the problem is the infinitely thin line the ball is being bounced from. Possible solution is to use a rotated rectangles to represent flippers instead, then do derotated AABB as the primary collision detection.